+++ .gitignore
@@ -14,3 +14,4 @@
 .externalNativeBuild
 .cxx
 local.properties
+client_secret_572756897592-lp22nas04q5ltvranu9jan0lr6qo2bc8.apps.googleusercontent.com.json
