+++ app/build.gradle
@@ -6,6 +6,10 @@ android {
     namespace 'com.example.musementfrontend'
     compileSdk 35
 
+    buildFeatures {
+        buildConfig true
+    }
+
     defaultConfig {
         applicationId "com.example.musementfrontend"
         minSdk 24
@@ -14,6 +18,7 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        buildConfigField "String", "GOOGLE_CLIENT_ID", "\"${GOOGLE_CLIENT_ID}\""
     }
 
     buildTypes {
@@ -38,11 +43,13 @@ dependencies {
     implementation libs.activity
     implementation libs.constraintlayout
     implementation libs.lombok
+    implementation 'com.google.android.gms:play-services-auth:20.0.1'
+    implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     testImplementation libs.junit
     androidTestImplementation libs.ext.junit
     androidTestImplementation libs.espresso.core
     implementation libs.glide
     annotationProcessor libs.compiler
     annotationProcessor 'org.projectlombok:lombok:1.18.30'
-
 }
