+++ app/src/main/java/com/example/musementfrontend/Client/APIClient.java
@@ -7,10 +7,9 @@ import retrofit2.converter.jackson.JacksonConverterFactory;
 
 public class APIClient {
 
-    private APIClient() {
-    }
-
     private static Retrofit retrofit = null;
+
+    private static final String BASE_URL = "http://10.0.2.2:8080";
 
     public static Retrofit getClient() {
         if (retrofit == null) {
@@ -18,7 +17,7 @@ public class APIClient {
             interceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
             OkHttpClient client = new OkHttpClient.Builder().addInterceptor(interceptor).build();
             retrofit = new Retrofit.Builder()
-                    .baseUrl("http://10.0.2.2:8080")
+                    .baseUrl(BASE_URL)
                     .addConverterFactory(JacksonConverterFactory.create())
                     .client(client)
                     .build();
