+++ app/src/main/java/com/example/musementfrontend/util/Util.java
@@ -35,12 +35,12 @@ import java.util.List;
 
 public class Util {
 
+    private static final String ICON_FILENAME = "musement_icon.png";
+
     static public void setIcon(AppCompatActivity activity){
         ImageView imageView = activity.findViewById(R.id.musement_icon);
 
-        String filename = "musement_icon.png";
-
-        try(InputStream inputStream = activity.getApplicationContext().getAssets().open(filename)){
+        try(InputStream inputStream = activity.getApplicationContext().getAssets().open(ICON_FILENAME)){
             Drawable drawable = Drawable.createFromStream(inputStream, null);
             imageView.setImageDrawable(drawable);
             imageView.setScaleType(ImageView.ScaleType.FIT_XY);
@@ -53,118 +53,5 @@ public class Util {
     static public EditText getCustomEditViewById(AppCompatActivity activity, int id){
         View view = activity.findViewById(id);
         return view.findViewById(R.id.edit_text_view);
-    }
-
-    static public void Init(AppCompatActivity activity){
-        InitHeader(activity);
-        InitMainMenu(activity);
-    }
-
-    static private void InitMainMenu(AppCompatActivity activity){
-        ConstraintLayout mainMenu = activity.findViewById(R.id.main_menu);
-        ImageButton feed = mainMenu.findViewById(R.id.feed);
-        ImageButton recommendation = mainMenu.findViewById(R.id.recommendation);
-        ImageButton invitation = mainMenu.findViewById(R.id.invitation);
-        ImageButton profile = mainMenu.findViewById(R.id.profile);
-
-        feed.setOnClickListener(Util::OnClickFeed);
-        recommendation.setOnClickListener(Util::OnClickRecommendation);
-        invitation.setOnClickListener(Util::OnClickInvitation);
-        profile.setOnClickListener(Util::OnClickProfile);
-
-    }
-
-    static private void InitHeader(AppCompatActivity activity){
-        ConstraintLayout header = activity.findViewById(R.id.header);
-        ImageButton bell = header.findViewById(R.id.bell);
-        ImageButton musement_icon = header.findViewById(R.id.musement_icon);
-
-        bell.setOnClickListener(Util::OnClickBell);
-        musement_icon.setOnClickListener(Util::OnClickMusementIcon);
-    }
-
-    static public void FillFeedConcert(AppCompatActivity activity,List<Concert> concerts){
-        ScrollView scroll = activity.findViewById(R.id.scroll);
-        ConstraintLayout layout = scroll.findViewById(R.id.feed_item);
-        LinearLayout feed = layout.findViewById(R.id.feed);
-
-        for(Concert concert : concerts){
-            View concertView = activity.getLayoutInflater().inflate(R.layout.concert_item, feed, false);
-            ImageButton concertImage = concertView.findViewById(R.id.concert);
-            Glide.with(activity)
-                    .load(concert.getImageUrl())
-                    .into(concertImage);
-            TextView artist = concertView.findViewById(R.id.artist);
-            TextView location = concertView.findViewById(R.id.location);
-            TextView date = concertView.findViewById(R.id.date);
-            artist.setText("Dima Bilan"); // change!!!
-            location.setText(concert.getLocation());
-            date.setText(concert.getDate().toString());
-            ViewGroup.MarginLayoutParams params = (ViewGroup.MarginLayoutParams) concertView.getLayoutParams();
-
-            params.setMargins(0, 0,0 , 80);
-            concertView.setLayoutParams(params);
-            feed.addView(concertView);
-        }
-    }
-
-    static public void OnClickBell(View view){
-        Context context = view.getContext();
-        if(context.getClass() == Notification.class){
-            return;
-        }
-        Intent intent = new Intent(context, Notification.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
-        context.startActivity(intent);
-    }
-
-    static public void OnClickMusementIcon(View view){
-        Context context = view.getContext();
-        if(context.getClass() == Feed.class){
-            return;
-        }
-        Intent intent = new Intent(context, Feed.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
-        context.startActivity(intent);
-    }
-
-    static private void OnClickFeed(View view){
-        Context context = view.getContext();
-        if(context.getClass() == Feed.class){
-            return;
-        }
-        Intent intent = new Intent(context, Feed.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
-        context.startActivity(intent);
-    }
-
-    static private void OnClickRecommendation(View view){
-        Context context = view.getContext();
-        if(context.getClass() == Recommendation.class){
-            return;
-        }
-        Intent intent = new Intent(context, Recommendation.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
-        context.startActivity(intent);
-    }
-
-    static private void OnClickInvitation(View view){
-        Context context = view.getContext();
-        if(context.getClass() == Invitation.class){
-            return;
-        }
-        Intent intent = new Intent(context, Invitation.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
-        context.startActivity(intent);
-    }
-
-    static private void OnClickProfile(View view){
-        Context context = view.getContext();
-        if(context.getClass() == Profile.class){
-            return;
-        }
-        Intent intent = new Intent(context, Profile.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
-        context.startActivity(intent);
     }
 }
