+++ gradle.properties
@@ -19,3 +19,4 @@ android.useAndroidX=true
 # resources declared in the library itself and none from the library's dependencies,
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
+GOOGLE_CLIENT_ID = $GOOGLE_CLIENT_ID
