+++ gradle/libs.versions.toml
@@ -2,6 +2,7 @@
 agp = "8.5.0"
 compiler = "4.16.0"
 converterJackson = "2.9.0"
+googleServices = "4.4.2"
 jacksonDatabind = "2.13.3"
 junit = "4.13.2"
 junitVersion = "1.2.1"
@@ -12,11 +13,13 @@ material = "1.12.0"
 activity = "1.10.0"
 constraintlayout = "2.2.0"
 lombok = "1.18.30"
+playServicesAuth = "21.3.0"
 
 [libraries]
 compiler = { module = "com.github.bumptech.glide:compiler", version.ref = "compiler" }
 converter-jackson = { module = "com.squareup.retrofit2:converter-jackson", version.ref = "converterJackson" }
 glide = { module = "com.github.bumptech.glide:glide", version.ref = "compiler" }
+google-services = { module = "com.google.gms:google-services", version.ref = "googleServices" }
 jackson-databind = { module = "com.fasterxml.jackson.core:jackson-databind", version.ref = "jacksonDatabind" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 ext-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
@@ -26,6 +29,7 @@ logging-interceptor = { module = "com.squareup.okhttp3:logging-interceptor", ver
 material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 activity = { group = "androidx.activity", name = "activity", version.ref = "activity" }
 constraintlayout = { group = "androidx.constraintlayout", name = "constraintlayout", version.ref = "constraintlayout" }
+play-services-auth = { module = "com.google.android.gms:play-services-auth", version.ref = "playServicesAuth" }
 retrofit = { module = "com.squareup.retrofit2:retrofit" }
 lombok = { group = "org.projectlombok", name = "lombok", version.ref = "lombok" }
